refactor(system): remove non-compliant adapter and branded signal header
- DELETE compat/syscalls_adapter.c - monolithic adapter sink removed
- DELETE include/ixland/ixland_signal.h - branded signal header removed
- Remove #include <signal.h> from include/ixland/ixland_syscalls.h
- Move public wrappers to canonical owner files:
  * fork, vfork -> kernel/fork.c
  * getpid, getppid, getpgrp, getpgid, setpgid, setsid, getsid -> kernel/task.c
  * sigaction, signal, kill, killpg, sigprocmask, sigpending, sigsuspend, raise, pause -> kernel/signal.c
  * ixland_cleanup -> kernel/init.c
- Fix exit() noreturn with pthread_exit
- Fix fs/select.c/h: rename to IX_FD_SETSIZE, ix_fd_set_t, IX_FD_ZERO, etc.
- Fix kernel/resource.c: add struct rlimit64 forward declaration
- Fix fs/path.c: C23 label declaration warning
- Fix kernel/net/network.c: size_t->uint32_t casts
- Fix fs/eventpoll.c: unused function warning

BUILD: GCC_TREAT_WARNINGS_AS_ERRORS=YES

+++ kernel/net/network.c
@@ -455,7 +455,7 @@ static ssize_t recv_impl(int sockfd, void *buf, size_t len, int flags) {
     dispatch_semaphore_t sem = dispatch_semaphore_create(0);
     __block ssize_t bytes_received = 0;
 
-    nw_connection_receive(sock->connection, 1, len,
+    nw_connection_receive(sock->connection, 1, (uint32_t)len,
                           ^(dispatch_data_t content, nw_content_context_t context, bool is_complete,
                             nw_error_t error) {
                               (void)context;
@@ -752,7 +752,7 @@ static ssize_t recvmsg_impl(int sockfd, struct msghdr *msg, int flags) {
     __block ssize_t bytes_received = 0;
     __block dispatch_data_t received_data = NULL;
 
-    nw_connection_receive(sock->connection, 1, total_len,
+    nw_connection_receive(sock->connection, 1, (uint32_t)total_len,
                           ^(dispatch_data_t content, nw_content_context_t context, bool is_complete, nw_error_t receive_error) {
                               if (receive_error) {
                                   bytes_received = -1;
